Content Model: Persist cache 1
